Added procedural Agent generation
