+++ Unity/Assets/Materials/Agent.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: fcad6d4e4f3ff767da5827489daf5016
-folderAsset: yes
-DefaultImporter:
+guid: 2921938675cf8e8668fc737e00c1be09
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
