+++ Unity/Assets/Scenes/TileWorld.unity
@@ -277,6 +277,23 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4d51bcb0270c5d7299e1001fd9ebb5c5, type: 3}
+--- !u!1 &1433068113 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 7522897998521695982, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
+  m_PrefabInstance: {fileID: 2103147849}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1433068114
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1433068113}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ca95cf342a6993cfcac0f612ba36bf69, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &1610169891
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -383,6 +400,10 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
+      propertyPath: agents
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
       propertyPath: mTiles
       value: 10
       objectReference: {fileID: 0}
@@ -403,17 +424,21 @@ PrefabInstance:
       value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
+      propertyPath: AgentPrefab
+      value: 
+      objectReference: {fileID: 6395324215237540279, guid: 635724beb6cc0e00e8aa743563626a85, type: 3}
+    - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
       propertyPath: tileMaterial
       value: 
-      objectReference: {fileID: 0}
+      objectReference: {fileID: 2100000, guid: 57a0ff389e7acf7829f07d00536a097f, type: 2}
     - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
       propertyPath: wallMaterial
       value: 
-      objectReference: {fileID: 0}
+      objectReference: {fileID: 2100000, guid: 81a1ad358e634d051a90e7d524b40eae, type: 2}
     - target: {fileID: 1492976209649619076, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
       propertyPath: groundMaterial
       value: 
-      objectReference: {fileID: 0}
+      objectReference: {fileID: 2100000, guid: 5b4cef496e45fddc69ff71f45f0451e9, type: 2}
     - target: {fileID: 7522897998521695982, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
       propertyPath: m_Name
       value: Enviroment
@@ -421,7 +446,10 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 7522897998521695982, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1433068114}
   m_SourcePrefab: {fileID: 100100000, guid: f56dc3bb783ec915a8404ac1940862a7, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
