+++ Unity/ProjectSettings/TagManager.asset
@@ -14,7 +14,7 @@ TagManager:
   - Sensor
   - Agent
   - Object
-  - 
+  - Obstacle
   - 
   - 
   -
